xcb: Avoid to pop the acquired screen from the entries
This prevent the acquisition thread to detect a resolution change.

Because the entry is removed from the entry list, a new entry will be
detected during the next call of iter_monitors(). This new entry will be
used to detect the resolution change, but the acquisition thread won't
be notified.